backend/dev-requirements.txt
appnope==0.1.0
astroid==2.2.5
backcall==0.1.0
cycler==0.10.0
Cython==0.29.15
decorator==4.4.0
Django==2.2.3
django-cors-headers==3.0.2
django-guardian==2.0.0
django-polymorphic==2.0.3
djangorestframework==3.9.4
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.14.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
matplotlib==3.2.0
mccabe==0.6.1
numpy==1.18.1
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycocotools==2.0.0
Pygments==2.4.2
pylint==2.3.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.1
scipy==1.4.1
six==1.12.0
sqlparse==0.3.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.43.0
traitlets==4.3.2
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.1
